Feature (SoftwareRenderer): Software rendering works with Vertex Array Objects lets gooooooooo

+++ JEngine3D/JEngine3D/Renderer/IRendererAPI.hpp
@@ -19,7 +19,7 @@ public:
   virtual void SetClearColor(const Color &color) = 0;
   virtual void Clear() = 0;
 
-  virtual void DrawIndexed(const IVertexArray &vertexArray, uint32_t indexCount = 0) = 0;
+  virtual void DrawIndexed(const IVertexArray &vertexArray, size_t indexCount = 0) = 0;
 };
 
 }// namespace JE
